* commit 'da343db3cadfa12cdd3b8856832f2ddef1870518': (28 commits)
  Add a parameter obs_wind_projected (default = on) [MAGP-912]
  Add a parameter obs_wind_projected (default = on) [MAGP-912]
  More typos fixed
  Fix typo in the name of a regression test
  fix typo in test name
  InputMatrix are not handled properly the wrap-around [MAGP-911]
  Small typo in regression tests [MAGP-891]
  Wind in synop plotting should be reprojected according to the projection.[MAGP-891] Fix in Polar and proj4 .. Added a full suite of regression  tests
  Tyop in odbviewver
  Fix regression test for horizontal bars [MAGP-889]
  Update referenec version to 2.30.0 for the regression suite
  Wind in synop plotting should be reprojected according to the projection [MAGP-891] Also implemented for proj4
  Adding support for horizontal bars [MAGP-889] Adding parameter graph_bar_orientation Adding regression tests
  Code clean-up - spaces and tabs
  scan-build - remove unused variable  MAGP-759
  SVG WIND - correct style for wind barbs  MAGP-517
  CMAKE - Replaced all message CMake functions with ecbuild_ones  MAGP-802
  LOGO - improve aspect ration of logos in Cairo driver   MAGP-570
  LOGO - first version of new logo in PostScript driver   MAGP-570
  LOGO - change colour to ECMWF_BLUE for new logo  MAGP-570
  ...

+++ src/decoders/GribSatelliteInterpretor.h
@@ -49,7 +49,7 @@ private:
     //! Copy constructor - No copy allowed
     GribSatelliteInterpretor(const GribSatelliteInterpretor&);
     //! Overloaded << operator to copy - No copy allowed
-    void AdjustBadlyEncodedGribs(int satId, int chanId, long &nx, long &ny, long &dx, long &dy, double &xp, double &yp, double &slon, long &functionCode) const;
+    void AdjustBadlyEncodedGribs(int satId, int chanId, long &nx, long &ny, long &dx, long &dy, long &xp, long &yp, double &slon, long &functionCode) const;
 
 
 // -- Friends
